A chord is now an ordered set
Previously, given an expression like "leftshift:l", we would send the key
events for LEFTSHIFT and L in an indeterminate order; this would result in
LEFTSHIFT being a no-op if it was sent second.

This commit replaces the previous simple type alias Chord with a struct. We
also perform a similar change for Sequence. A practical impact of this is
that we have been forced to rewrite client code to use the interface
methods rather than direct access.